enemy shooting and health. going to work on movements of enemies

+++ TarotPlatformer/Assets/Code/System/Enemy/EnemyManager.cs
@@ -6,7 +6,7 @@ public class EnemyManager {
 
     public static float speed = 3f;
     public static float jump_speed = 500f;
-    public static int health = 100;
+    public static float health = 100f;
     public static float damage = 10f;
     public static float atk_speed = 10f;
